Rank Defactor
Defactored Rank and should fast update after game.
