+++ Assets/Scripts/RankManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1edcb2c6fe4abd34d909010cd910ff31
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
